requirements.txt
django==4.2.11
django-phonenumber-field[phonenumbers]